Refactored jacobianResidual function, added a Profile class, updated recorder function
The jacobianResidual function is refactored, moved the constraint application and dynamic terms to their own functions. Added a Profile class for creating lines, sin arches (full and half). The old recorder functions is updated which can dump displacement and internal force readings into files. linkerRodsNew and rodPairsNew are renamed to regular version without the New, and the corresponding regular files are renamed to linkerRodsOld and rodPairsOld